fix: erro pasta não encontrada
