--- NotificaSpring/notifica/src/main/java/br/com/dbc/vemser/notifica/controller/UsuarioController.java
@@ -1,4 +1,0 @@
-package br.com.dbc.vemser.notifica.controller;
-
-public class UsuarioController {
-}
