--- NotificaSpring/notifica/src/main/java/br/com/dbc/vemser/notifica/entity/Localizacao.java
@@ -1,4 +1,0 @@
-package models;
-
-public record Localizacao(double latitude, double longitude) {
-}
