--- NotificaSpring/notifica/src/main/java/br/com/dbc/vemser/notifica/exceptions/RegraDeNegocioException.java
@@ -1,8 +1,0 @@
-package br.com.dbc.vemser.pessoaapi.exceptions;
-
-public class RegraDeNegocioException extends Exception {
-
-    public RegraDeNegocioException(String message) {
-        super(message);
-    }
-}
