--- NotificaSpring/notifica/src/main/java/br/com/dbc/vemser/notifica/repository/UsuarioRepository.java
@@ -1,4 +1,0 @@
-package br.com.dbc.vemser.notifica.repository;
-
-public class UsuarioRepository {
-}
