--- NotificaSpring/notifica/src/main/java/br/com/dbc/vemser/notifica/service/UsuarioService.java
@@ -1,4 +1,0 @@
-package br.com.dbc.vemser.notifica.service;
-
-public class UsuarioService {
-}
