+++ NotificaSpring/notifica2/src/main/java/br/com/dbc/vemser/notifica/NotificaApplication.java
@@ -1,0 +1,13 @@
+package br.com.dbc.vemser.notifica;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NotificaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(NotificaApplication.class, args);
+	}
+
+}
