+++ NotificaSpring/notifica2/src/main/java/br/com/dbc/vemser/notifica/controller/DenunciaController.java
@@ -1,0 +1,4 @@
+package br.com.dbc.vemser.notifica.controller;
+
+public class DenunciaController {
+}
