+++ NotificaSpring/notifica2/src/main/java/br/com/dbc/vemser/notifica/entity/Localizacao.java
@@ -1,0 +1,4 @@
+package br.com.dbc.vemser.notifica.entity;
+
+public record Localizacao(double latitude, double longitude) {
+}
