+++ NotificaSpring/notifica2/src/main/java/br/com/dbc/vemser/notifica/exceptions/RegraDeNegocioException.java
@@ -1,0 +1,8 @@
+package br.com.dbc.vemser.notifica.exceptions;
+
+public class RegraDeNegocioException extends Exception {
+
+    public RegraDeNegocioException(String message) {
+        super(message);
+    }
+}
