+++ NotificaSpring/notifica2/src/main/java/br/com/dbc/vemser/notifica/repository/DenunciaRepository.java
@@ -1,0 +1,4 @@
+package br.com.dbc.vemser.notifica.repository;
+
+public class DenunciaRepository {
+}
