+++ NotificaSpring/notifica2/src/main/java/br/com/dbc/vemser/notifica/service/DenunciaService.java
@@ -1,0 +1,4 @@
+package br.com.dbc.vemser.notifica.service;
+
+public class DenunciaService {
+}
